requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.0
motor==3.3.2
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0 